fix: split UsageRecord type and UsageCollectorAPI into src/types
@mulmocast/types builds from src/types/ via per-file symlinks under
types/src/ — src/utils/ is intentionally not included. Having
MulmoStudioContext reference UsageCollector directly from src/utils/
broke `yarn build` in the types package with TS2307.

- Move `UsageRecord` and a new `UsageCollectorAPI` interface to
  src/types/usage.ts (browser-safe, types-only).
- `src/utils/usage_collector.ts` now imports the type and the class
  implements `UsageCollectorAPI`. Re-export keeps the public API stable.
- `MulmoStudioContext.usageCollector?` switches to `UsageCollectorAPI`
  so the types package compiles without seeing the concrete class.
- Symlink usage.ts into types/src/ and re-export from
  src/types/index.ts so @mulmocast/types consumers can see it.

Both `yarn build` (root) and `cd types && yarn build` now succeed.
