+++ types/src/usage.ts
@@ -1,0 +1,1 @@
+../../src/types/usage.ts
